requirements.txt
selenium==4.10.0
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
openai==1.3.0
schedule==1.2.0
fake-useragent==1.1.3
instaloader==4.9.5